Add RefreshToken DelegateSubdomainZone API calls to Defang Service
RefreshToken API call would be used to refresh the defang token as well
as checking the customer login status.

The DelegateSubdomainZone API call allows the customer to create
subdomain delegation in defang managed domains.

Modified GrpcClient and ByocClient to allow BYOC to login to a defang
server to give access to ByocClient the TenantID and defang Token, which
could be used by CD to make API calls to Defang fabric to delegate the
subdomain to customer's own route53 subdomain zone.