Use AWS cli export credential as the first provider in the chain
This would make sure defang cli would behave exactly the same in how to
look up the credentials as the aws cli, since aws cli have a different
behavior for sso logged in users, which does not save the credentials to
the standard location the go sdk check, and if there is a wrong/expired
credential saved with the same profile name, the aws cli would still
succeed but defang cli would fail which could be very confusing for
customers.

Also updated the hint for if the aws credentials expired.
Do not defaults to aws provider when there is only AWS_REGION
environment variable is detected.