use NewLoader to set and verify composeFilePath
Use the compose-go library to find the default path if no path is
passed instead of replicating their logic in our codebase.

Once a path is obtained, we can derive the current working
directory.

If we are using default paths, we will test if there are
potentially conflicting default compose files: compose.yaml _and_
docker-compose.yaml.