update to change DeleteSecret to DeleteConfig.
add support for show all

fix to show error on config not founc.

fix typo

make sure aws errors are not surfaced to user, update with alternative message

update so path is not part of the names from list and get functions. Update list to use GetParametersByPath for better efficiency