Alias compose commands at root (#618)
* move compose commands to a new file

* refactor compose command setup

to use factory functions instead of global structs. this will make
it easier to create different instances of these command structs and
hang them in different places on the tree.

* alias compose commands at root

* move logs command into compose commands, and alias on root

* Hide root `start`, `stop` and `restart` commands for now

They aren't implemented quite correctly at the moment.

* use ps to list services

docker compose ls actually lists projects, but because we currently only
support one project in the playground, we're going to alias ls and ps
for now. when we add multi-project support to the playground, we can
distinguish them from one another.