Added to the flake.nix and add new scripts
Added to the flake.nix that will make a .git if there
is none and install the pre-push hook if it is not
already installed.

scripts/git-hooks/pre-push is used to prevent direct
pushes to a target branch. If the current
branch is branchProtection, the push will be rejected.
branchProtection is my branch name, for testing purposes.